refactor: extract service features
